fixed tck_io bug and a few other things

+++ src/bonndit/utils/tck_io.py
@@ -74,12 +74,12 @@ class Tck:
                     f.write(b'0' * (10 - len(str(self.length))) + bytes(str(self.length), encoding='utf-8'))
 
         with open(self.file_path, 'ab') as f:
-            f.write(np.ndarray.tobytes(np.array([np.inf, np.inf, np.inf]), dtype='<f4'))
+            f.write(np.ndarray.tobytes(np.array([np.inf, np.inf, np.inf], dtype='<f4')))
             print('File with %s streamlines saved in %s' % (self.length, self.file_path))
         if self.feature:
             for feat in self.feature.keys():
                 with open(self.feature[feat]['path'], 'ab') as f:
-                    f.write(np.ndarray.tobytes(np.array([np.inf])))
+                    f.write(np.ndarray.tobytes(np.array([np.inf], dtype='<f4')))
                 print('File with feature %s saved in %s' % (feat, self.feature[feat]['path']))
 
     def read(self):
